+++ packages/twenty-server/src/emails/components/BaseEmail.tsx
@@ -1,8 +1,8 @@
 import * as React from 'react';
 import { Container, Html } from '@react-email/components';
 
-import { BaseHead } from 'src/workspace/emails/components/BaseHead';
-import { Logo } from 'src/workspace/emails/components/Logo';
+import { BaseHead } from 'src/emails/components/BaseHead';
+import { Logo } from 'src/emails/components/Logo';
 
 export const BaseEmail = ({ children }) => {
   return (
