+++ packages/twenty-server/src/emails/components/BaseHead.tsx
