+++ packages/twenty-server/src/emails/components/HighlightedText.tsx
