+++ packages/twenty-server/src/emails/components/Logo.tsx
