+++ packages/twenty-server/src/emails/components/MainText.tsx
